# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Bank API",
        "description": "Bank OpenApi description",
        "contact": {
            "email": "bogdantegiero@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=L5-Swagger"
        }
    },
    "paths": {
        "/api/payment-accounts": {
            "post": {
                "tags": [
                    "Accounts"
                ],
                "summary": "Создаёт новый платёжный счёт",
                "operationId": "bf3327884d9309daca9598fa35905e2f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "client_id": {
                                        "description": "id банка",
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "1"
                                    },
                                    "password": {
                                        "description": "id клиента",
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Account created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "account has been created"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Database error: ..."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "field": {
                                                "type": "string",
                                                "example": "email"
                                            },
                                            "message": {
                                                "type": "string",
                                                "example": "The email has already been taken."
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/credit-accounts": {
            "post": {
                "tags": [
                    "Accounts"
                ],
                "summary": "Создание кредитного счёта",
                "operationId": "b88aed1c96a8899161e683f14f5c0a25",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "client_id": {
                                        "description": "ID of the client",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "bank_id": {
                                        "description": "ID of the bank",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "employee_id": {
                                        "description": "ID of the employee who processed the credit",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "payment_account_id": {
                                        "description": "ID of the client's payment account",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "start_credit": {
                                        "description": "Credit start date",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-03-15"
                                    },
                                    "end_credit": {
                                        "description": "Credit end date",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-03-15"
                                    },
                                    "month_credit": {
                                        "description": "Credit term in months",
                                        "type": "integer",
                                        "example": 12
                                    },
                                    "credit_sum": {
                                        "description": "Credit amount",
                                        "type": "number",
                                        "format": "float",
                                        "example": 10000
                                    },
                                    "month_pay": {
                                        "description": "Monthly payment amount",
                                        "type": "number",
                                        "format": "float",
                                        "example": 1000
                                    },
                                    "percent_credit": {
                                        "description": "Month rate",
                                        "type": "number",
                                        "format": "float",
                                        "example": 10
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Credit account created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Credit account has been created"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "field": {
                                                "type": "string",
                                                "example": "client_id"
                                            },
                                            "message": {
                                                "type": "string",
                                                "example": "The selected client id is invalid."
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Database error: ..."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}